rename dialogs to match file names
